feat: relationship between scan filters
