+++ ui/components/scans/index.ts
@@ -2,3 +2,4 @@ export * from "./auto-refresh";
 export * from "./link-to-findings-from-scan";
 export * from "./no-providers-added";
 export * from "./no-providers-connected";
+export * from "./scans-filters";
