+++ ui/hooks/index.ts
@@ -1,5 +1,6 @@
 export * from "./use-credentials-form";
 export * from "./use-form-server-errors";
 export * from "./use-local-storage";
+export * from "./use-related-filters";
 export * from "./use-sidebar";
 export * from "./use-store";
